bug fix in length data type

+++ Raydreams.Common/Model/FileWrappers.cs
@@ -66,7 +66,7 @@ namespace Raydreams.Common.Model
 
         /// <summary>The byte length of the original decoded data</summary>
         [JsonProperty( "length" )]
-        public int Filelength { get; set; } = 0;
+        public long Filelength { get; set; } = 0;
 
         /// <summary>Quick check the object has everything to be valid</summary>
         /// <remarks>ContentType is optional since it can fallback to checking the filename or assume its a default.</remarks>
